removed dependencies on Colt + other cleanup

+++ symbolization/src/main/java/geovista/symbolization/SimplePaletteImpl.java
@@ -70,7 +70,7 @@ public class SimplePaletteImpl implements Palette{
     }
     
     private int recalculateAnchorPosition(int oldPosition, int newSize){
-        return Math.round(((float)oldPosition / (float)noOfInputColors) * (float)newSize);
+        return Math.round(((float)oldPosition / (float)noOfInputColors) * newSize);
     }
     
     public void setName(String s){
